Moved LLM query recommendation to experimental category. Moved demo of LLM recommendations to its own demo with it clearly being marked as experimental. Made sure query logging is disabled unless user explicitly enables it.
