+++ src/pd_explain/explainers/fedex_explainer.py
@@ -6,8 +6,8 @@ import re
 
 from fedex_generator.Operations.Filter import Filter
 from pd_explain.llm_integrations import ExplanationReasoning
-from pd_explain.query_recommenders.query_logger import QueryLogger
-from pd_explain.query_recommenders.query_score_functions import score_queries
+from pd_explain.experimental.query_recommenders import QueryLogger
+from pd_explain.experimental.query_recommenders import score_queries
 
 
 class FedexExplainer(ExplainerInterface):
