--- src/pd_explain/query_recommenders/consts.py
@@ -1,3 +1,0 @@
-DOT_ENV_PD_EXPLAIN_LOG_QUERIES = "PD_EXPLAIN_LOG_QUERIES"
-DOT_ENV_PD_EXPLAIN_LOG_FILE_LOCATION = "PD_EXPLAIN_LOG_FILE_LOCATION"
-PD_EXPLAIN_DEFAULT_LOG_FILE_LOCATION = "pd_explain.log.csv"
